Custom Exception and Response Object Created
* Created Custom NotFoundException, BadRequestException and ValidationException
* Integrated those exceptions in return type of command Handlers
* Created BaseCommandResponse class for returning response.

+++ Hr.LeaveManagement.Application/Features/LeaveRequests/Handlers/Commands/DeleteLeaveRequestCommandHandler.cs
@@ -1,5 +1,7 @@
-﻿using Hr.LeaveManagement.Application.Features.LeaveRequests.Requests.Commands;
+﻿using Hr.LeaveManagement.Application.Exceptions;
+using Hr.LeaveManagement.Application.Features.LeaveRequests.Requests.Commands;
 using Hr.LeaveManagement.Application.Persistence.Contracts;
+using Hr.LeaveManagement.Domain;
 using MediatR;
 
 namespace Hr.LeaveManagement.Application.Features.LeaveRequests.Handlers.Commands
@@ -15,6 +17,10 @@ namespace Hr.LeaveManagement.Application.Features.LeaveRequests.Handlers.Command
         public async Task<Unit> Handle(DeleteLeaveRequestCommand request, CancellationToken cancellationToken)
         {
             var leaveRequest = await _leaveRequestRepository.Get(request.Id);
+
+            if (leaveRequest == null)
+                throw new NotFoundException(nameof(LeaveRequest), request.Id);
+
             await _leaveRequestRepository.Delete(leaveRequest);
             return Unit.Value;
         }
